# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Aurigraph V11 Platform API
  description: |
    High-performance blockchain platform APIs for 2M+ TPS
    
    ## Overview
    Aurigraph V11 is a next-generation blockchain platform built with Java 21, Quarkus, and GraalVM.
    It provides high-throughput transaction processing, quantum-resistant cryptography, cross-chain bridging,
    and AI-driven optimization capabilities.
    
    ## Features
    - **Ultra-High Throughput**: 2M+ transactions per second
    - **Quantum-Resistant Security**: Post-quantum cryptography implementation
    - **Cross-Chain Bridge**: Interoperability with Ethereum, Solana, and other chains
    - **AI Optimization**: Machine learning-driven consensus and performance optimization
    - **HMS Integration**: Healthcare Management System integration
    - **Real-time Monitoring**: Comprehensive performance and health monitoring
    
    ## Performance Targets
    - **TPS**: 2,000,000+ transactions per second
    - **Latency**: Sub-millisecond transaction processing
    - **Startup**: <1 second native binary startup
    - **Memory**: <256MB native binary footprint
    
    ## Technology Stack
    - **Runtime**: Java 21 with Virtual Threads
    - **Framework**: Quarkus 3.26.2
    - **Native**: GraalVM native compilation
    - **Transport**: HTTP/2 with TLS 1.3, gRPC
    - **Consensus**: HyperRAFT++ algorithm
    - **Cryptography**: CRYSTALS-Kyber/Dilithium post-quantum algorithms
    
  version: 11.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /gateway/auth/token:
    post:
      tags:
        - API Gateway
      summary: Generate authentication token
      description: Generate an authentication token for API access
      operationId: generateToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    TokenRequest:
      type: object
      properties:
        clientId:
          type: string
          description: Client identifier
        clientSecret:
          type: string
          description: Client secret
        scopes:
          type: array
          items:
            type: string
          description: Requested scopes
      required:
        - clientId
        - clientSecret
        - scopes
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Access token
        token_type:
          type: string
          description: Token type (Bearer)
        expires_in:
          type: integer
          description: Token expiration in seconds
        expires_at:
          type: integer
          format: int64
          description: Token expiration timestamp
        scope:
          type: array
          items:
            type: string
          description: Granted scopes

    # Common Schemas
...